google log in working and log out working, can colse userinfo and email and send to django
